requirements.txt
python-docx>=1.1.0            # Для чтения файлов docx
natasha>=1.6.0                # Русскоязычный NLP (NER, токенизация)
spacy>=3.7.4                  # Универсальная NLP библиотека (NER, кореференция)
pymorphy3>=1.2.0              # Морфологический анализ

spacy-langdetect>=0.1.2       # Определение языка текста

# Работа с LLM и суммаризацией
transformers>=4.40.0          # HuggingFace Transformers для работы с LLM
sentence-transformers>=2.7.0  # Эмбеддинги для RAG
langchain>=0.2.0              # Удобная реализация RAG-пайплайна
faiss-cpu>=1.8.0              # Векторный поиск для RAG

# Дополнительно (опционально)
fastcoref>=2.1.6              # Эффективная кореференция (опционально, рекомендуется)

# Служебные библиотеки
tqdm>=4.66.4                  # Прогресс-бар для отслеживания выполнения задач
loguru>=0.7.2                 # Удобное логирование
numpy<2.0

